Testing the Story drive module Struggling painfully to come up with a decent name for the class that will run the story of the game. Decided (susceptible to change) that the story will be written in INK, a software used to make text based - choice based - branching game; story will be exported as JSON file from the software and then read using the opensource jar/library made by some very kind people and hosted on github under bladecoder/ink ,for reading an INK.json.
Initial tests have been successful. Was able to run a simple choice
based text adventure of 80Days.
Choices are displayed as text and the user gives integer input
corresponding to the index value of the choice starting from 0.

Must learn to implement a way to make the branches alter the
attributes of the player class based on choices and the present running
branch.

Haven't yet figured out how to save and load the branch files.

@neilmadhava   ...    I am really excited about this now.
With the ascii text art and visual art, our little game has started
looking like a proper retro text based game.
With the use of Ink for implementing a story, our game is finally
staring to take a decent shape.

Always looking for story suggestions, improvemnets and probablly even a
reason to scrath our cheesy story altogether for a better one.